Added a Fancyzoom-based file chooser.
Each textarea on a page gets a link that, when clicked, pops up a list of the files that have been uploaded to the site. A Markdown-formatted image is then inserted into the textarea at the caret position and replaces any selected text.

I'd like to tidy it up a bit at some point in the future, but I'm going to use it for a while to see how I find it.