My server setup is now fairly complicated, inasmuch as the Django site is cached with StaticGenerator and served with Nginx.
Nginx serves directories regardless of whether the URL ends in a slash or not.  So I make sure it does by including a rewrite in the config.  The trouble is, this causes a clash with Django URLs that *don't* end in a slash, so I'm changing the only one I have on the site to end in a slash.